+++ CHANGES.md
@@ -1,4 +1,5 @@
 # 6.5.4 2023-??-??
+ - 2023-06-19 Fixed dynamic field type 'Webservice' admin dialog test field being read-only and not usable.
  - 2023-06-02 S/MIME: Increased size of column 'email_address' of table 'smime_keys'.
 
 # 6.5.3 2023-06-14